physics: Huge optimisation by using grid

+++ engine/src/lib2d/window.hpp
@@ -68,11 +68,19 @@ namespace frog::lib2d::gx
 
         void draw_colored( const texture& tex, int u, int v,
                 int cut_width, int cut_height, int x, int y,
-                int tex_width, int tex_height, std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a );
+                int tex_width, int tex_height,
+                std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a );
+
+        // void draw_colored_rotated( const texture& tex, int u, int v,
+        //                         int cut_width, int cut_height, int x, int y,
+        //                         int tex_width, int tex_height,
+        //                         std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a,
+        //                         int pivot_x, int pivot_y, float angle,
+        //                         bool flipped = false );
 
         void draw_colored_rotated( const texture& tex, int u, int v,
-                                int cut_width, int cut_height, int x, int y,
-                                int tex_width, int tex_height,
+                                int cut_width, int cut_height, float x, float y,
+                                float tex_width, float tex_height,
                                 std::uint8_t r, std::uint8_t g, std::uint8_t b, std::uint8_t a,
                                 int pivot_x, int pivot_y, float angle,
                                 bool flipped = false );
